Finish

+++ src/Controller/ActivityController.php
@@ -3,8 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Activity;
+use App\Entity\Coach;
 use App\Form\ActivityType;
 use App\Repository\ActivityRepository;
+use App\Repository\CalendarRepository;
 use App\Service\ControlUpload;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -13,6 +15,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use DateTime;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
  * @Route("/activity")
@@ -40,6 +43,7 @@ class ActivityController extends AbstractController
 
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $activity->setOwner($this->getUser());
             $activity->setUpdateAt(new \DateTime('now'));
             if ($activity->getCreatedAt() === null) {
                 $activity->setCreatedAt(new \DateTime('now'));
@@ -76,8 +80,35 @@ class ActivityController extends AbstractController
      */
     public function show(Activity $activity): Response
     {
+        $user = $this->getUser();
+
+        $calendars = $activity->getActivityRDV();
+        
+        $event = [];
+        try {
+            $event[] = [
+                'id' => $calendars->getId(),
+                'start' => $calendars->getStart()->format('Y-m-d H:i'),
+                'end' => $calendars->getEnd()->format('Y-m-d H:i'),
+                'title' => $calendars->getTitle(),
+                'description' => $calendars->getDescription(),
+                'backgroundColor' => $calendars->getBackgroundColor(),
+                'borderColor' => $calendars->getBorderColor(),
+                'textColor' => $calendars->getTextColor(),
+            ];
+        } catch (\Throwable $th) {
+            //throw $th;
+        }
+        
+        $data = json_encode($event);
+
+        $coaches = $activity->getCoaches();
+
         return $this->render('activity/show.html.twig', [
             'activity' => $activity,
+            'coaches' => $coaches,
+            'data' => $data,
+            'user' => $user
         ]);
     }
 
@@ -86,6 +117,11 @@ class ActivityController extends AbstractController
      */
     public function edit(Request $request, Activity $activity, EntityManagerInterface $entityManager): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if (!($this->getUser() == $activity->getOwner())) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw new AccessDeniedException('Only the owner can edit the program!');
+        }
         $form = $this->createForm(ActivityType::class, $activity);
         $form->handleRequest($request);
 
@@ -106,6 +142,11 @@ class ActivityController extends AbstractController
      */
     public function delete(Request $request, Activity $activity, EntityManagerInterface $entityManager): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if (!($this->getUser() == $activity->getOwner())) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw new AccessDeniedException('Only the owner can edit the program!');
+        }
         if ($this->isCsrfTokenValid('delete' . $activity->getId(), $request->request->get('_token'))) {
             $entityManager->remove($activity);
             $entityManager->flush();
